Finish decoupling GraphModel from specific type implementation
