+++ src/gui-qml/qml/DissolveMain.qml
@@ -169,7 +169,7 @@ ApplicationWindow {
         }
         Item {
             id: exampleGraphTab
-            GraphModel {
+            ExampleGraphModel {
                 id: graphModel
             }
